+++ src/Stormancer.Plugins/Analytics/Stormancer.Server.Plugins.Analytics/AnalyticsExtensions.cs
@@ -8,8 +8,15 @@ using System.Threading.Tasks;
 
 namespace Stormancer
 {
+    /// <summary>
+    /// Extension methods for <see cref="ISceneHost"/>.
+    /// </summary>
     public static class AnalyticsExtensions
     {
+        /// <summary>
+        /// Adds Analytics functionalities to the scene.
+        /// </summary>
+        /// <param name="scene"></param>
         public static void AddAnalytics(this ISceneHost scene)
         {
             scene.Metadata[AnalyticsPlugin.METADATA_KEY] = "enabled";