+++ src/Stormancer.Plugins/Api/Stormancer.Server.Plugins.Api/IControllerFactory.cs
@@ -30,9 +30,15 @@ using Stormancer;
 using System.Linq.Expressions;
 using Stormancer.Diagnostics;
 using System.IO;
+using System.Diagnostics;
 
 namespace Stormancer.Server.Plugins.API
 {
+
+
+
+
+
     /// <summary>
     /// Constants related to APIs.
     /// </summary>
@@ -57,6 +63,7 @@ namespace Stormancer.Server.Plugins.API
         /// <summary>
         /// The API can be called by scene hosts.
         /// </summary>
+        [Obsolete("Using packet based messaging for S2S is deprecated. Decorate with S2SApiAttribute to use the new Pipeline based S2S communication API.")]
         Scene2Scene
     }
 
@@ -79,6 +86,9 @@ namespace Stormancer.Server.Plugins.API
     /// <summary>
     /// Attributes marking a method as a remotely callable API.
     /// </summary>
+    /// <remarks>
+    /// For S2S API, use <see cref="S2SApiAttribute"/>.
+    /// </remarks>
     public class ApiAttribute : Attribute
     {
 
@@ -147,6 +157,7 @@ namespace Stormancer.Server.Plugins.API
             }
         }
 
+
         private async Task ExecuteConnected(IScenePeerClient client)
         {
             using (var scope = _scene.DependencyResolver.CreateChild(Constants.ApiRequestTag))
@@ -164,6 +175,32 @@ namespace Stormancer.Server.Plugins.API
                 var controller = scope.Resolve<T>();
                 controller.Peer = args.Peer;
                 await controller.OnDisconnected(args);
+            }
+        }
+
+        private Task ExecuteSceneStarting()
+        {
+            using (var scope = _scene.DependencyResolver.CreateChild(Constants.ApiRequestTag))
+            {
+                var controller = scope.Resolve<T>();
+                var handlers = _scene.DependencyResolver.Resolve<IEnumerable<IApiHandler>>();
+                var logger = _scene.DependencyResolver.Resolve<ILogger>();
+                return handlers.RunEventHandler(h => h.SceneStarting(_scene, controller), ex => logger.Log(LogLevel.Error, "api", $"An error occured on an handler while executing {nameof(IApiHandler.SceneStarting)}", ex));
+
+
+            }
+        }
+
+        private Task ExecuteShuttingDown(ShutdownArgs args)
+        {
+            using (var scope = _scene.DependencyResolver.CreateChild(Constants.ApiRequestTag))
+            {
+                var controller = scope.Resolve<T>();
+                var handlers = _scene.DependencyResolver.Resolve<IEnumerable<IApiHandler>>();
+                var logger = _scene.DependencyResolver.Resolve<ILogger>();
+                return handlers.RunEventHandler(h => h.SceneShuttingDown(_scene, controller), ex => logger.Log(LogLevel.Error, "api", $"An error occured on an handler while executing {nameof(IApiHandler.SceneShuttingDown)}", ex));
+
+
             }
         }
 
@@ -220,6 +257,64 @@ namespace Stormancer.Server.Plugins.API
             }
         }
 
+
+        private async Task ExecuteS2SRequest(ApiCallContext<IS2SRequestContext> ctx, Func<T, IS2SRequestContext, IDependencyResolver, Task> action)
+        {
+            using (var scope = _scene.DependencyResolver.CreateChild(Constants.ApiRequestTag, ctx.Builder))
+            {
+                var controller = scope.Resolve<T>();
+                try
+                {
+                    if (controller == null)
+                    {
+                        throw new InvalidOperationException("The controller could not be found. Make sure it has been properly registered in the dependency manager.");
+                    }
+
+                    controller.CancellationToken = ctx.Context.CancellationToken;
+
+                    await action(controller, ctx.Context, scope);
+                }
+                catch (ClientException)
+                {
+                    throw;
+                }
+                catch (OperationCanceledException)
+                {
+                    throw new ClientException("canceled");
+                }
+                catch (AggregateException ex)
+                {
+                    var clientEx = ex.InnerExceptions.FirstOrDefault(ex => ex is ClientException);
+                    if (clientEx == null || ex.InnerExceptions.Count > 1)
+                    {
+                        if (controller == null || !await controller.HandleException(new ApiExceptionContext(ctx.Route, ex, ctx.Context)))
+                        {
+                            scope.Resolve<ILogger>().Log(LogLevel.Error, ctx.Route, $"An exception occurred while executing action '{ctx.Route}' in controller '{typeof(T).Name}'.", ex);
+
+                        }
+                    }
+
+                    if (clientEx != null)
+                    {
+                        throw clientEx;
+                    }
+                    else
+                    {
+                        throw;
+                    }
+
+                }
+                catch (Exception ex)
+                {
+                    if (controller == null || !await controller.HandleException(new ApiExceptionContext(ctx.Route, ex, ctx.Context)))
+                    {
+                        scope.Resolve<ILogger>().Log(LogLevel.Error, ctx.Route, $"An exception occurred while executing action '{ctx.Route}' in controller '{typeof(T).Name}'.", ex);
+
+                    }
+                    throw;
+                }
+            }
+        }
         private async Task ExecuteRpcAction(ApiCallContext<RequestContext<IScenePeerClient>> ctx, Func<T, RequestContext<IScenePeerClient>, IDependencyResolver, Task> action)
         {
             using (var scope = _scene.DependencyResolver.CreateChild(Constants.ApiRequestTag, ctx.Builder))
@@ -253,7 +348,7 @@ namespace Stormancer.Server.Plugins.API
                         if (controller == null || !await controller.HandleException(new ApiExceptionContext(ctx.Route, ex, ctx.Context)))
                         {
                             scope.Resolve<ILogger>().Log(LogLevel.Error, ctx.Route, $"An exception occurred while executing action '{ctx.Route}' in controller '{typeof(T).Name}'.", ex);
-                            
+
                         }
                     }
 
@@ -272,7 +367,7 @@ namespace Stormancer.Server.Plugins.API
                     if (controller == null || !await controller.HandleException(new ApiExceptionContext(ctx.Route, ex, ctx.Context)))
                     {
                         scope.Resolve<ILogger>().Log(LogLevel.Error, ctx.Route, $"An exception occurred while executing action '{ctx.Route}' in controller '{typeof(T).Name}'.", ex);
-                       
+
                     }
                     throw;
                 }
@@ -300,7 +395,7 @@ namespace Stormancer.Server.Plugins.API
                 {
                     throw;
                 }
-                catch(OperationCanceledException)
+                catch (OperationCanceledException)
                 {
                     throw new ClientException("canceled");
                 }
@@ -322,6 +417,9 @@ namespace Stormancer.Server.Plugins.API
             _scene.ConnectionRejected.Add(p => ExecuteConnectionRejected(p));
             _scene.Connected.Add(p => ExecuteConnected(p));
             _scene.Disconnected.Add(args => ExecuteDisconnected(args));
+            _scene.Shuttingdown.Add(args => ExecuteShuttingDown(args));
+            _scene.Starting.Add(_ => ExecuteSceneStarting());
+
             foreach (var method in type.GetMethods())
             {
                 if (TryAddRoute(type, method))
@@ -418,13 +516,53 @@ namespace Stormancer.Server.Plugins.API
 
         }
 
+
+
         private bool TryAddRoute(Type controllerType, MethodInfo method)
         {
-            var attr = method.GetCustomAttribute<ApiAttribute>();
-            if (attr == null)
+            var apiAttr = method.GetCustomAttribute<ApiAttribute>();
+            if (apiAttr != null)
             {
-                return false;
+                return TryAddRoute(controllerType, method, apiAttr);
             }
+
+            var s2sAttr = method.GetCustomAttribute<S2SApiAttribute>();
+
+            if (s2sAttr != null)
+            {
+                return TryAddRoute(controllerType, method, s2sAttr);
+            }
+            return false;
+
+        }
+
+        private bool TryAddRoute(Type controllerType, MethodInfo method, S2SApiAttribute attr)
+        {
+            var route = attr.Route ?? GetProcedureName(controllerType, method, false);
+            var returnType = method.ReturnType;
+
+            var a = CreateExecuteActionFunctionS2S(method);
+            _scene.AddS2SRequestHandler(route, request =>
+            {
+
+                var initialApiCallCtx = new ApiCallContext<IS2SRequestContext>(request, method, route);
+
+                Func<ApiCallContext<IS2SRequestContext>, Task> next = ctx => ExecuteS2SRequest(ctx, (c, pctx, r) => a(c, pctx, r));
+
+                foreach (var handler in _scene.DependencyResolver.Resolve<Func<IEnumerable<IApiHandler>>>()())
+                {
+                    next = WrapWithHandler(handler.RunS2S, next);
+                }
+
+                return next(initialApiCallCtx);
+            });
+
+            return true;
+
+        }
+
+        private bool TryAddRoute(Type controllerType, MethodInfo method, ApiAttribute attr)
+        {
             var route = attr.Route ?? GetProcedureName(controllerType, method, false);
             var returnType = method.ReturnType;
             if (attr.Type == ApiType.FireForget && returnType != typeof(Task) && returnType != typeof(void))
@@ -548,9 +686,11 @@ namespace Stormancer.Server.Plugins.API
 
 
 
+
         private static Func<T, TRq, IDependencyResolver, Task> CreateExecuteActionFunction<TRq>(MethodInfo method)
         {
             var returnType = method.ReturnType;
+
             if (returnType == typeof(void) || returnType == typeof(Task))
             {
                 return ApiHelpers.CreateExecuteActionFunctionVoid<TRq>(method);
@@ -586,6 +726,77 @@ namespace Stormancer.Server.Plugins.API
             }
         }
 
+        private static Func<T, IS2SRequestContext, IDependencyResolver, Task> CreateExecuteActionFunctionS2S(MethodInfo method)
+        {
+            var returnType = method.ReturnType;
+
+            var controller = Expression.Parameter(typeof(T), "controller");
+            var ctx = Expression.Parameter(typeof(IS2SRequestContext), "ctx");
+            var resolver = Expression.Parameter(typeof(IDependencyResolver), "resolver");
+
+            if (returnType == typeof(void) || returnType == typeof(Task))
+            {
+
+
+                // Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>>
+                var readArgsFunction = Expression.Constant(
+                    ApiHelpers.CreateReadArgumentsS2S(method.GetParameters())
+                    , typeof(Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>>)
+                );
+
+                //public static Func<T, List<object?>, IDependencyResolver, Task> CreateExecuteActionFunctionS2S<TReturn>(MethodInfo method)
+                var executeFunction = Expression.Constant(
+                    ApiHelpers.CreateExecuteActionFunctionVoidS2S(method),
+                    typeof(Func<T, List<object?>, IDependencyResolver, Task>)
+                );
+
+                //Func<T, IS2SRequestContext, IDependencyResolver,
+                //    Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>>,
+                //    Func<T, List<object?>, IDependencyResolver, Task>,
+                //    Task
+                //>
+                var wrapperMethod = typeof(ApiHelpers).GetRuntimeMethodExt("ExecuteActionVoidS2S", p => true);
+
+                return Expression.Lambda<Func<T, IS2SRequestContext, IDependencyResolver, Task>>(Expression.Call(wrapperMethod, controller, ctx, resolver, readArgsFunction, executeFunction), controller, ctx, resolver).Compile();
+            }
+            else
+            {
+                if (returnType.IsGenericType && returnType.GetGenericTypeDefinition() == typeof(Task<>))
+                {
+                    returnType = returnType.GetGenericArguments()[0];
+
+                }
+
+                // Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>>
+                var readArgsFunction = Expression.Constant(
+                    ApiHelpers.CreateReadArgumentsS2S(method.GetParameters())
+                    , typeof(Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>>));
+
+                //public static Func<T, List<object?>, IDependencyResolver, Task<TReturn>> CreateExecuteActionFunctionS2S<TReturn>(MethodInfo method)
+                var executeFunction = Expression.Constant(
+                    typeof(ApiHelpers).GetRuntimeMethodExt("CreateExecuteActionFunctionS2S", p => true).MakeGenericMethod(returnType).Invoke(null, new[] { method }),
+                    typeof(Func<,,,>).MakeGenericType(typeof(T), typeof(List<object?>), typeof(IDependencyResolver), typeof(Task<>).MakeGenericType(returnType))
+                    );
+
+                //public static Func<TRq, TReturn, IDependencyResolver> CreateWriteResultLambda<TRq, TReturn>()
+                var sendResultFunction = Expression.Constant(
+                    typeof(ApiHelpers).GetRuntimeMethodExt("CreateWriteResultLambda", p => true).MakeGenericMethod(typeof(IS2SRequestContext), returnType).Invoke(null, new object[] { }),
+                    typeof(Func<,,,>).MakeGenericType(typeof(IS2SRequestContext), returnType, typeof(IDependencyResolver), typeof(Task))
+                    );
+
+                //public static async Task Task ExecuteActionAndSendResultS2S<TReturn>(T controller, IS2SRequestContext ctx, IDependencyResolver resolver,
+                //  Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>> readArguments,
+                //  Func< T, List<object?>, IDependencyResolver, Task < TReturn >> executeControllerActionFunction,
+                //  Func<IS2SRequestContext, TReturn, IDependencyResolver, Task> sendResultAction)
+                var wrapperMethod = typeof(ApiHelpers).GetRuntimeMethodExt("ExecuteActionAndSendResultS2S", p => true).MakeGenericMethod(returnType);
+
+
+
+                return Expression.Lambda<Func<T, IS2SRequestContext, IDependencyResolver, Task>>(Expression.Call(wrapperMethod, controller, ctx, resolver, readArgsFunction, executeFunction, sendResultFunction), controller, ctx, resolver).Compile();
+            }
+        }
+
+
 
 
         private static class ApiHelpers
@@ -594,6 +805,64 @@ namespace Stormancer.Server.Plugins.API
             public static void ReadObject<TData>(RequestContext<IScenePeer> ctx, out TData output, IDependencyResolver resolver) => output = resolver.Resolve<ISerializer>().Deserialize<TData>(ctx.InputStream);
             public static void ReadObject<TData>(Packet<IScenePeerClient> ctx, out TData output, IDependencyResolver resolver) => output = ctx.ReadObject<TData>();
             public static void ReadObject<TData>(RequestContext<IScenePeerClient> ctx, out TData output, IDependencyResolver resolver) => output = ctx.ReadObject<TData>();
+
+            public static async Task<ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver>> ReadObject<TData>(ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver> tuple)
+            {
+                var (rq, args, resolver) = tuple;
+                var serializer = resolver.Resolve<ISerializer>();
+                args.Add(await serializer.DeserializeAsync<TData>(rq.Reader, rq.CancellationToken));
+                return tuple;
+            }
+            public static Task<ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver>> ReadObjectCtx(ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver> tuple)
+            {
+                var (rq, args, _) = tuple;
+                args.Add(rq);
+                return Task.FromResult(tuple);
+            }
+
+            public static Task<ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver>> CreateArgsReadSeedS2S(IS2SRequestContext ctx, IDependencyResolver resolver) => Task.FromResult((ctx, new List<object?>(), resolver));
+
+            public static Task<ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver>> ReadNextS2SArgument<TData>(Task<ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver>> task) => task.ContinueWith(t => ReadObject<TData>(task.Result)).Unwrap();
+
+            public static Task<ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver>> ReadNextS2SArgumentCtx(Task<ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver>> task) => task.ContinueWith(t => ReadObjectCtx(task.Result)).Unwrap();
+
+            public static Task<List<object?>> CompleteReadS2SArgs(Task<ValueTuple<IS2SRequestContext, List<object?>, IDependencyResolver>> task) => task.ContinueWith(t => t.Result.Item2);
+
+
+            public static Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>> CreateReadArgumentsS2S(ParameterInfo[] args)
+            {
+
+                var ctx = Expression.Parameter(typeof(IS2SRequestContext), "ctx");
+                var resolver = Expression.Parameter(typeof(IDependencyResolver), "resolver");
+
+                var expressions = new List<Expression>();
+                var expr = Expression.Call(typeof(ApiHelpers).GetRuntimeMethodExt(nameof(CreateArgsReadSeedS2S), _ => true), ctx, resolver);
+
+                //Call Read recurcively to chain the continueWith and read all arguments.
+                foreach (var parameter in args)
+                {
+                    if (parameter.ParameterType == typeof(IS2SRequestContext))
+                    {
+                        expr = Expression.Call(typeof(ApiHelpers).GetRuntimeMethodExt(nameof(ReadNextS2SArgumentCtx), _ => true), expr);
+                    }
+                    else
+                    {
+                        expr = Expression.Call(typeof(ApiHelpers).GetRuntimeMethodExt(nameof(ReadNextS2SArgument), _ => true).MakeGenericMethod(parameter.ParameterType), expr);
+                    }
+                }
+
+                expr = Expression.Call(typeof(ApiHelpers).GetRuntimeMethodExt(nameof(CompleteReadS2SArgs), _ => true), expr);
+
+                expressions.Add(expr);
+
+                return Expression.Lambda<Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>>>(Expression.Block(expressions), ctx, resolver).Compile();
+            }
+
+            public static Task WriteResult<TData>(IS2SRequestContext ctx, TData value, IDependencyResolver resolver)
+            {
+                var serializer = resolver.Resolve<ISerializer>();
+                return serializer.SerializeAsync(value, ctx.Writer, ctx.CancellationToken);
+            }
 
             public static Task WriteResult<TData>(RequestContext<IScenePeerClient> ctx, TData value, IDependencyResolver resolver)
             {
@@ -606,6 +875,20 @@ namespace Stormancer.Server.Plugins.API
                     return Task.CompletedTask;
                 }
             }
+
+            public static Task WriteResult<TData>(RequestContext<IScenePeer> ctx, TData value, IDependencyResolver resolver)
+            {
+                if (!ctx.CancellationToken.IsCancellationRequested)
+                {
+                    var serializer = resolver.Resolve<ISerializer>();
+                    return ctx.SendValue(s => serializer.Serialize(value, s));
+                }
+                else
+                {
+                    return Task.CompletedTask;
+                }
+            }
+
 
             public static async Task WriteResultAsyncEnumerable<TData>(RequestContext<IScenePeerClient> ctx, IAsyncEnumerable<TData> value, IDependencyResolver resolver)
             {
@@ -621,20 +904,20 @@ namespace Stormancer.Server.Plugins.API
                     }
                 }
             }
-            public static Task WriteResult<TData>(RequestContext<IScenePeer> ctx, TData value, IDependencyResolver resolver)
+
+            public static async Task WriteResultAsyncEnumerable<TData>(IS2SRequestContext ctx, IAsyncEnumerable<TData> enumerable, IDependencyResolver resolver)
             {
-                if (!ctx.CancellationToken.IsCancellationRequested)
+                var serializer = resolver.Resolve<ISerializer>();
+                await foreach (var result in enumerable)
                 {
-                    var serializer = resolver.Resolve<ISerializer>();
-                    return ctx.SendValue(s => serializer.Serialize(value, s));
-                }
-                else
-                {
-                    return Task.CompletedTask;
+                    await serializer.SerializeAsync(result, ctx.Writer, ctx.CancellationToken);
                 }
             }
 
-            public static async Task WriteResulttAsyncEnumerable<TData>(RequestContext<IScenePeer> ctx, IAsyncEnumerable<TData> values, IDependencyResolver resolver)
+
+
+
+            public static async Task WriteResultAsyncEnumerable<TData>(RequestContext<IScenePeer> ctx, IAsyncEnumerable<TData> values, IDependencyResolver resolver)
             {
                 await foreach (var value in values)
                 {
@@ -650,11 +933,76 @@ namespace Stormancer.Server.Plugins.API
                 }
             }
 
+            public static async Task ExecuteActionAndSendResultS2S<TReturn>(T controller, IS2SRequestContext ctx, IDependencyResolver resolver,
+                Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>> readArguments,
+                Func<T, List<object?>, IDependencyResolver, Task<TReturn>> executeControllerActionFunction,
+                Func<IS2SRequestContext, TReturn, IDependencyResolver, Task> sendResultAction)
+            {
+                var args = await readArguments(ctx, resolver);
+                var result = await executeControllerActionFunction(controller, args, resolver);
+                await sendResultAction(ctx, result, resolver);
+            }
+
+            public static async Task ExecuteActionVoidS2S(T controller, IS2SRequestContext ctx, IDependencyResolver resolver, Func<IS2SRequestContext, IDependencyResolver, Task<List<object?>>> readArgsFunction, Func<T, List<object?>, IDependencyResolver, Task> executeFunction)
+            {
+                var args = await readArgsFunction(ctx, resolver);
+                await executeFunction(controller, args, resolver);
+            }
+
             public static async Task ExecuteActionAndSendResult<TRq, TReturn>(T controller, TRq ctx, IDependencyResolver resolver, Func<T, TRq, IDependencyResolver, Task<TReturn>> executeControllerActionFunction, Func<TRq, TReturn, IDependencyResolver, Task> sendResultAction)
             {
                 var result = await executeControllerActionFunction(controller, ctx, resolver);
                 await sendResultAction(ctx, result, resolver);
             }
+
+            public static Func<T, List<object?>, IDependencyResolver, Task> CreateExecuteActionFunctionVoidS2S(MethodInfo method)
+            {
+                var controller = Expression.Parameter(typeof(T), "controller");
+                var args = Expression.Parameter(typeof(List<object?>), "args");
+                var resolver = Expression.Parameter(typeof(IDependencyResolver), "resolver");
+
+                var expressions = new List<Expression>();
+                var parameters = AddReadParametersExpressions(expressions, args, method.GetParameters());
+
+                var callExpression = Expression.Call(controller, method, parameters);
+                expressions.Add(callExpression);
+                if (method.ReturnType == typeof(void))
+                {
+
+                    expressions.Add(Expression.Constant(Task.CompletedTask));
+                }
+
+                return Expression.Lambda<Func<T, List<object?>, IDependencyResolver, Task>>(Expression.Block(parameters, expressions), controller, args, resolver).Compile();
+
+            }
+
+            public static Func<T, List<object?>, IDependencyResolver, Task<TReturn>> CreateExecuteActionFunctionS2S<TReturn>(MethodInfo method)
+            {
+                var controller = Expression.Parameter(typeof(T), "controller");
+                var args = Expression.Parameter(typeof(List<object?>), "args");
+                var resolver = Expression.Parameter(typeof(IDependencyResolver), "resolver");
+
+                var expressions = new List<Expression>();
+                var parameters = AddReadParametersExpressions(expressions, args, method.GetParameters());
+
+                var callExpression = Expression.Call(controller, method, parameters);
+                expressions.Add(callExpression);
+                if (method.ReturnType.IsGenericType && method.ReturnType.GetGenericTypeDefinition() == typeof(Task<>))
+                {
+                    expressions.Add(callExpression);
+                }
+                else
+                {
+                    var taskFromResultMethod = typeof(Task).GetRuntimeMethodExt("FromResult", p => true);
+                    Debug.Assert(taskFromResultMethod != null);
+                    expressions.Add(Expression.Call(taskFromResultMethod.MakeGenericMethod(typeof(TReturn)), callExpression));
+                }
+
+                return Expression.Lambda<Func<T, List<object?>, IDependencyResolver, Task<TReturn>>>(Expression.Block(parameters, expressions), controller, args, resolver).Compile();
+
+            }
+
+
 
             /// <summary>
             /// Creates a lambda expression that extract the parameters from the request then call the action and returns a Task 
@@ -687,6 +1035,7 @@ namespace Stormancer.Server.Plugins.API
 
 
 
+
             public static Func<T, TRq, IDependencyResolver, Task<TReturn>> CreateExecuteActionFunction<TRq, TReturn>(MethodInfo method)
             {
                 var controller = Expression.Parameter(typeof(T), "controller");
@@ -704,7 +1053,9 @@ namespace Stormancer.Server.Plugins.API
                 }
                 else
                 {
-                    expressions.Add(Expression.Call(typeof(Task).GetRuntimeMethodExt("FromResult", p => true).MakeGenericMethod(typeof(TReturn)), callExpression));
+                    var taskFromResultMethod = typeof(Task).GetRuntimeMethodExt("FromResult", p => true);
+                    Debug.Assert(taskFromResultMethod != null);
+                    expressions.Add(Expression.Call(taskFromResultMethod.MakeGenericMethod(typeof(TReturn)), callExpression));
                 }
 
                 return Expression.Lambda<Func<T, TRq, IDependencyResolver, Task<TReturn>>>(Expression.Block(parameters, expressions), controller, ctx, resolver).Compile();
@@ -734,12 +1085,34 @@ namespace Stormancer.Server.Plugins.API
                     else
                     {
                         var readObjectMethod = typeof(ApiHelpers).GetRuntimeMethodExt("ReadObject", p => p[0].ParameterType == typeof(TRq));//Select good ReadObject overload
-
+                        Debug.Assert(readObjectMethod != null);
                         block.Add(Expression.Call(readObjectMethod.MakeGenericMethod(parameterInfo.ParameterType), ctx, variables[i], resolver));
                     }
                 }
                 return variables;
+
             }
+
+            public static ParameterExpression[] AddReadParametersExpressions(List<Expression> block, ParameterExpression args, ParameterInfo[] actionParameters)
+            {
+
+                var length = actionParameters.Length;
+                var variables = new ParameterExpression[actionParameters.Length];
+
+                for (int i = 0; i < length; i++)
+                {
+                    var parameterInfo = actionParameters[i];
+                    var variable = Expression.Variable(parameterInfo.ParameterType, parameterInfo.Name);
+
+                    variables[i] = variable;
+
+                    block.Add(Expression.Assign(variables[i], Expression.Convert(Expression.MakeIndex(args, typeof(List<object?>).GetProperty("Item"), new[] { Expression.Constant(i) }), variables[i].Type)));
+
+                }
+                return variables;
+
+            }
+
 
             //Action<TRq, TReturn, IDependencyResolver>
             public static Func<TRq, TReturn, IDependencyResolver, Task> CreateWriteResultLambda<TRq, TReturn>()
@@ -754,14 +1127,18 @@ namespace Stormancer.Server.Plugins.API
                 if (returnType.IsGenericType && returnType.GetGenericTypeDefinition() == typeof(IAsyncEnumerable<>))
                 {
                     var writeResultMethod = typeof(ApiHelpers).GetRuntimeMethodExt("WriteResultAsyncEnumerable", p => p[0].ParameterType == typeof(TRq));
+                    Debug.Assert(writeResultMethod != null);
                     return Expression.Lambda<Func<TRq, TReturn, IDependencyResolver, Task>>(Expression.Call(writeResultMethod.MakeGenericMethod(returnType.GetGenericArguments()[0]), ctx, data, resolver), ctx, data, resolver).Compile();
                 }
                 else
                 {
                     var writeResultMethod = typeof(ApiHelpers).GetRuntimeMethodExt("WriteResult", p => p[0].ParameterType == typeof(TRq));
+                    Debug.Assert(writeResultMethod != null);
                     return Expression.Lambda<Func<TRq, TReturn, IDependencyResolver, Task>>(Expression.Call(writeResultMethod.MakeGenericMethod(returnType), ctx, data, resolver), ctx, data, resolver).Compile();
                 }
             }
+
+
         }
     }
     internal static class RuntimeMethodExtensions
@@ -778,7 +1155,9 @@ namespace Stormancer.Server.Plugins.API
                               select ele);
 
             // Maybe check if we have more than 1? Or not?
-            return potentials.FirstOrDefault();
+            var method = potentials.FirstOrDefault();
+            Debug.Assert(method != null);
+            return method;
         }
         public static MethodInfo GetRuntimeMethodExt(this Type type, string name, Func<ParameterInfo[], bool> parametersFilter)
         {
@@ -791,7 +1170,9 @@ namespace Stormancer.Server.Plugins.API
                               select ele);
 
             // Maybe check if we have more than 1? Or not?
-            return potentials.FirstOrDefault();
+            var method = potentials.FirstOrDefault();
+            Debug.Assert(method != null);
+            return method;
         }
     }