+++ src/Stormancer.Plugins/Notification/Stormancer.Server.Plugins.Notification/Changelog.rst
@@ -9,6 +9,11 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+2.0.2
+-----
+Changed
+*******
+- Use Stormancer.Abstraction.Server 4.2.0.1-pre
 
 2.0.1
 -----